+++ src/OrderLinesList/OrdersLinesSearchConfig.js
@@ -33,9 +33,9 @@ export const queryTemplate = generateQueryTemplate(indexes);
 export const getKeywordQuery = query => indexes.reduce(
   (acc, sIndex) => {
     if (acc) {
-      return `${acc} or ${sIndex}="${query}*"`;
+      return `${acc} or ${sIndex}=="*${query}*"`;
     } else {
-      return `${sIndex}="${query}*"`;
+      return `${sIndex}=="*${query}*"`;
     }
   },
   '',