+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 
+* Add substring searching for orders and order lines search options. Refs UIOR-623
 * "Orders: Can edit Order" doesn't allow me to view orders. Refs UIOR-625
 * Activate "re-encumber' toggle by default when creating order. Refs UIOR-622
 * Error message: Budget Expense Class not found. Refs UIOR-621.